requirements/conda.txt
numpy
scipy
psutil
textx
pyyaml
diffpy.srfit
diffpy.srreal
diffpy.structure
